requirements.txt
ackermann-msgs==2.0.2
action-msgs==2.0.3
action-tutorials-interfaces==0.33.9
action-tutorials-py==0.33.9
actionlib-msgs==5.3.6
actuator-msgs==0.0.1
ament-clang-format==0.17.4
ament-clang-tidy==0.17.4
ament-cmake-google-benchmark==2.5.5
ament-cmake-test==2.5.5
ament-copyright==0.17.4
ament-cppcheck==0.17.4
ament-cpplint==0.17.4
ament-flake8==0.17.4
ament-index-python==1.8.2
ament-lint==0.17.4
ament-lint-cmake==0.17.4
ament-package==0.16.5
ament-pep257==0.17.4
ament-uncrustify==0.17.4
ament-xmllint==0.17.4
angles==1.16.1
annotated-doc==0.0.4
anyio==4.12.1
builtin-interfaces==2.0.3
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
composition-interfaces==2.0.3
control-msgs==5.7.0
controller-manager==4.42.2
controller-manager-msgs==4.42.2
cuda-bindings==13.0.3
cuda-pathfinder==1.2.2
cv-bridge==4.1.0
demo-nodes-py==0.33.9
diagnostic-msgs==5.3.6
diagnostic-updater==4.2.6
domain-coordinator==0.12.0
example-interfaces==0.12.0
examples-rclpy-executors==0.19.7
examples-rclpy-minimal-action-client==0.19.7
examples-rclpy-minimal-action-server==0.19.7
examples-rclpy-minimal-client==0.19.7
examples-rclpy-minimal-publisher==0.19.7
examples-rclpy-minimal-service==0.19.7
examples-rclpy-minimal-subscriber==0.19.7
exceptiongroup==1.3.1
filelock==3.24.3
fsspec==2026.2.0
gazebo-msgs==3.8.0
generate-parameter-library-py==0.6.0
geometry-msgs==5.3.6
gps-msgs==2.1.1
graph-msgs==0.2.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.4.1
idna==3.11
image-geometry==4.1.0
interactive-markers==2.5.5
Jinja2==3.1.6
joint-state-publisher==2.4.0
joint-state-publisher-gui==2.4.0
laser-geometry==2.7.2
launch==3.4.10
launch-param-builder==0.1.1
launch-pytest==3.4.10
launch-ros==0.26.11
launch-testing==3.4.10
launch-testing-ros==0.26.11
launch-xml==3.4.10
launch-yaml==3.4.10
lifecycle-msgs==2.0.3
logging-demo==0.33.9
map-msgs==2.4.1
markdown-it-py==4.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
message-filters==4.11.10
moveit==0.1.4
moveit-configs-utils==2.14.1
moveit-msgs==2.7.1
moveit-task-constructor-msgs==0.1.4
mpmath==1.3.0
nav-msgs==5.3.6
networkx==3.4.2
numpy==2.2.6
nvidia-cublas==13.1.0.3
nvidia-cuda-cupti==13.0.85
nvidia-cuda-nvrtc==13.0.88
nvidia-cuda-runtime==13.0.96
nvidia-cudnn-cu13==9.15.1.9
nvidia-cufft==12.0.0.61
nvidia-cufile==1.15.1.6
nvidia-curand==10.4.0.35
nvidia-cusolver==12.0.4.66
nvidia-cusparse==12.6.3.3
nvidia-cusparselt-cu13==0.8.0
nvidia-nccl-cu13==2.28.9
nvidia-nvjitlink==13.0.88
nvidia-nvshmem-cu13==3.4.5
nvidia-nvtx==13.0.85
object-recognition-msgs==2.0.0
octomap-msgs==2.0.1
osrf-pycommon==2.1.7
packaging @ file:///home/task_176104877067765/conda-bld/packaging_1761049113113/work
pal-statistics==2.7.0
pal-statistics-msgs==2.7.0
pcl-msgs==1.0.0
pendulum-msgs==0.33.9
pillow==12.1.1
protobuf==7.34.0
Pygments==2.19.2
python-qt-binding==2.2.2
PyYAML==6.0.3
qt-dotgraph==2.7.5
qt-gui==2.7.5
qt-gui-cpp==2.7.5
qt-gui-py-common==2.7.5
quality-of-service-demo-py==0.33.9
rcl-interfaces==2.0.3
rclpy==7.1.9
rcutils==6.7.5
regex==2026.1.15
requests==2.32.5
resource-retriever==3.4.4
rich==14.3.2
rmw-dds-common==3.1.1
ros-gz-bridge==1.0.18
ros-gz-interfaces==1.0.18
ros-gz-sim==1.0.18
ros2action==0.32.8
ros2bag==0.26.9
ros2bag-mcap-cli==0.26.9
ros2bag-sqlite3-cli==0.26.9
ros2cli==0.32.8
ros2component==0.32.8
ros2controlcli==4.42.2
ros2doctor==0.32.8
ros2interface==0.32.8
ros2launch==0.26.11
ros2lifecycle==0.32.8
ros2multicast==0.32.8
ros2node==0.32.8
ros2param==0.32.8
ros2pkg==0.32.8
ros2plugin==5.4.4
ros2run==0.32.8
ros2service==0.32.8
ros2test==0.6.0
ros2topic==0.32.8
rosbag2-interfaces==0.26.9
rosbag2-py==0.26.9
rosgraph-msgs==2.0.3
rosidl-adapter==4.6.7
rosidl-cli==4.6.7
rosidl-cmake==4.6.7
rosidl-generator-c==4.6.7
rosidl-generator-cpp==4.6.7
rosidl-generator-py==0.22.2
rosidl-generator-type-description==4.6.7
rosidl-parser==4.6.7
rosidl-pycommon==4.6.7
rosidl-runtime-py==0.13.1
rosidl-typesupport-c==3.2.2
rosidl-typesupport-cpp==3.2.2
rosidl-typesupport-fastrtps-c==3.6.3
rosidl-typesupport-fastrtps-cpp==3.6.3
rosidl-typesupport-introspection-c==4.6.7
rosidl-typesupport-introspection-cpp==4.6.7
rpyutils==0.4.2
rqt-action==2.2.1
rqt-bag==1.5.6
rqt-bag-plugins==1.5.6
rqt-console==2.2.2
rqt-graph==1.5.6
rqt-gui==1.6.3
rqt-gui-py==1.6.3
rqt-msg==1.5.2
rqt-plot==1.4.5
rqt-publisher==1.7.3
rqt-py-common==1.6.3
rqt-py-console==1.2.3
rqt-reconfigure==1.6.3
rqt-service-caller==1.2.2
rqt-shell==1.2.3
rqt-srv==1.2.3
rqt-topic==1.7.5
safetensors==0.7.0
sensor-msgs==5.3.6
sensor-msgs-py==5.3.6
sentencepiece==0.2.1
service-msgs==2.0.3
shape-msgs==5.3.6
shellingham==1.5.4
simulation-interfaces==1.2.0
srdfdom==2.0.7
sros2==0.13.5
statistics-msgs==2.0.3
std-msgs==5.3.6
std-srvs==5.3.6
stereo-msgs==5.3.6
sympy==1.14.0
teleop-twist-keyboard==2.4.1
tf2-geometry-msgs==0.36.19
tf2-kdl==0.36.19
tf2-msgs==0.36.19
tf2-py==0.36.19
tf2-ros-py==0.36.19
tf2-sensor-msgs==0.36.19
tf2-tools==0.36.19
theora-image-transport==4.0.6
tokenizers==0.22.2
topic-monitor==0.33.9
topic-tools==1.3.3
topic-tools-interfaces==1.3.3
torch==2.10.0+cu130
torchvision==0.25.0+cu130
tqdm==4.67.3
trajectory-msgs==5.3.6
transformers==5.2.0
triton==3.6.0
turtlesim==1.8.3
type-description-interfaces==2.0.3
typer==0.24.0
typer-slim==0.24.0
typing_extensions==4.15.0
unique-identifier-msgs==2.5.0
urdfdom-py==1.2.1
urllib3==2.6.3
vision-msgs==4.1.1
visualization-msgs==5.3.6
xacro==2.1.1
